Order_Service/requirements.txt
ruamel.yaml==0.16.12
xmltodict==0.12.0
